+++ src/lib.rs
@@ -50,8 +50,7 @@ extern crate serde_json;
 #[cfg(feature = "ffi")]
 mod ffi;
 
-mod keys;
-pub use keys::{derive_key_ecdh_1pu, derive_key_ecdh_es, KeyAlg, KeyEntry, LocalKey};
+pub mod kms;
 
 mod protect;
 pub use protect::{generate_raw_store_key, PassKey, StoreKeyMethod};
